refactor: moved more stuff around :)
