+++ src/typedal/fields.py
@@ -2,27 +2,249 @@
 This file contains available Field types.
 """
 
+from __future__ import annotations
+
 import ast
+import contextlib
 import datetime as dt
 import decimal
-import typing
+import types
+import typing as t
 import uuid
 
+import pydal
 from pydal.helpers.classes import SQLCustomType
 from pydal.objects import Table
-from typing_extensions import Unpack
 
-from .core import TypeDAL, TypedField, TypedTable
-from .types import FieldSettings
+from .core import TypeDAL
+from .types import (
+    Expression,
+    Field,
+    FieldSettings,
+    Query,
+    T,
+    T_annotation,
+    T_MetaInstance,
+    T_subclass,
+    T_Value,
+    Validator,
+)
 
-T = typing.TypeVar("T", bound=typing.Any)
+if t.TYPE_CHECKING:
+    # will be imported for real later:
+    from .tables import TypedTable
 
 
 ## general
 
 
+class TypedField(Expression, t.Generic[T_Value]):  # pragma: no cover
+    """
+    Typed version of pydal.Field, which will be converted to a normal Field in the background.
+    """
+
+    # will be set by .bind on db.define
+    name = ""
+    _db: t.Optional[pydal.DAL] = None
+    _rname: t.Optional[str] = None
+    _table: t.Optional[Table] = None
+    _field: t.Optional[Field] = None
+
+    _type: T_annotation
+    kwargs: t.Any
+
+    requires: Validator | t.Iterable[Validator]
+
+    # NOTE: for the logic of converting a TypedField into a pydal Field, see TypeDAL._to_field
+
+    def __init__(
+        self,
+        _type: t.Type[T_Value] | types.UnionType = str,  # type: ignore
+        /,
+        **settings: t.Unpack[FieldSettings],
+    ) -> None:
+        """
+        Typed version of pydal.Field, which will be converted to a normal Field in the background.
+
+        Provide the Python type for this field as the first positional argument
+        and any other settings to Field() as keyword parameters.
+        """
+        self._type = _type
+        self.kwargs = settings
+        # super().__init__()
+
+    @t.overload
+    def __get__(self, instance: T_MetaInstance, owner: t.Type[T_MetaInstance]) -> T_Value:  # pragma: no cover
+        """
+        row.field -> (actual data).
+        """
+
+    @t.overload
+    def __get__(self, instance: None, owner: "t.Type[TypedTable]") -> "TypedField[T_Value]":  # pragma: no cover
+        """
+        Table.field -> Field.
+        """
+
+    def __get__(
+        self,
+        instance: T_MetaInstance | None,
+        owner: t.Type[T_MetaInstance],
+    ) -> t.Union[T_Value, "TypedField[T_Value]"]:
+        """
+        Since this class is a Descriptor field, \
+            it returns something else depending on if it's called on a class or instance.
+
+        (this is mostly for mypy/typing)
+        """
+        if instance:
+            # this is only reached in a very specific case:
+            # an instance of the object was created with a specific set of fields selected (excluding the current one)
+            # in that case, no value was stored in the owner -> return None (since the field was not selected)
+            return t.cast(T_Value, None)  # cast as T_Value so mypy understands it for selected fields
+        else:
+            # getting as class -> return actual field so pydal understands it when using in query etc.
+            return t.cast(TypedField[T_Value], self._field)  # pretend it's still typed for IDE support
+
+    def __str__(self) -> str:
+        """
+        String representation of a Typed Field.
+
+        If `type` is set explicitly (e.g. TypedField(str, type="text")), that type is used: `TypedField.text`,
+        otherwise the type annotation is used (e.g. TypedField(str) -> TypedField.str)
+        """
+        return str(self._field) if self._field else ""
+
+    def __repr__(self) -> str:
+        """
+        More detailed string representation of a Typed Field.
+
+        Uses __str__ and adds the provided extra options (kwargs) in the representation.
+        """
+        string_value = self.__str__()
+
+        if "type" in self.kwargs:
+            # manual type in kwargs supplied
+            typename = self.kwargs["type"]
+        elif issubclass(type, type(self._type)):
+            # normal type, str.__name__ = 'str'
+            typename = getattr(self._type, "__name__", str(self._type))
+        elif t_args := t.get_args(self._type):
+            # list[str] -> 'str'
+            typename = t_args[0].__name__
+        else:  # pragma: no cover
+            # fallback - something else, may not even happen, I'm not sure
+            typename = self._type
+
+        string_value = f"TypedField[{typename}].{string_value}" if string_value else f"TypedField[{typename}]"
+
+        kw = self.kwargs.copy()
+        kw.pop("type", None)
+        return f"<{string_value} with options {kw}>"
+
+    def _to_field(self, extra_kwargs: t.MutableMapping[str, t.Any]) -> t.Optional[str]:
+        """
+        Convert a Typed Field instance to a pydal.Field.
+
+        Actual logic in TypeDAL._to_field but this function creates the pydal type name and updates the kwarg settings.
+        """
+        other_kwargs = self.kwargs.copy()
+        extra_kwargs.update(other_kwargs)  # <- modifies and overwrites the default kwargs with user-specified ones
+        return extra_kwargs.pop("type", False) or TableDefinitionBuilder.annotation_to_pydal_fieldtype(
+            self._type,
+            extra_kwargs,
+        )
+
+    def bind(self, field: pydal.objects.Field, table: pydal.objects.Table) -> None:
+        """
+        Bind the right db/table/field info to this class, so queries can be made using `Class.field == ...`.
+        """
+        self._table = table
+        self._field = field
+
+    def __getattr__(self, key: str) -> t.Any:
+        """
+        If the regular getattribute does not work, try to get info from the related Field.
+        """
+        with contextlib.suppress(AttributeError):
+            return super().__getattribute__(key)
+
+        # try on actual field:
+        return getattr(self._field, key)
+
+    def __eq__(self, other: t.Any) -> Query:
+        """
+        Performing == on a Field will result in a Query.
+        """
+        return t.cast(Query, self._field == other)
+
+    def __ne__(self, other: t.Any) -> Query:
+        """
+        Performing != on a Field will result in a Query.
+        """
+        return t.cast(Query, self._field != other)
+
+    def __gt__(self, other: t.Any) -> Query:
+        """
+        Performing > on a Field will result in a Query.
+        """
+        return t.cast(Query, self._field > other)
+
+    def __lt__(self, other: t.Any) -> Query:
+        """
+        Performing < on a Field will result in a Query.
+        """
+        return t.cast(Query, self._field < other)
+
+    def __ge__(self, other: t.Any) -> Query:
+        """
+        Performing >= on a Field will result in a Query.
+        """
+        return t.cast(Query, self._field >= other)
+
+    def __le__(self, other: t.Any) -> Query:
+        """
+        Performing <= on a Field will result in a Query.
+        """
+        return t.cast(Query, self._field <= other)
+
+    def __hash__(self) -> int:
+        """
+        Shadow Field.__hash__.
+        """
+        return hash(self._field)
+
+    def __invert__(self) -> Expression:
+        """
+        Performing ~ on a Field will result in an Expression.
+        """
+        if not self._field:  # pragma: no cover
+            raise ValueError("Unbound Field can not be inverted!")
+
+        return t.cast(Expression, ~self._field)
+
+    def lower(self) -> Expression:
+        """
+        For string-fields: compare lowercased values.
+        """
+        if not self._field:  # pragma: no cover
+            raise ValueError("Unbound Field can not be lowered!")
+
+        return t.cast(Expression, self._field.lower())
+
+
+def is_typed_field(cls: t.Any) -> t.TypeGuard["TypedField[t.Any]"]:
+    """
+    Is `cls` an instance or subclass of TypedField?
+
+    Deprecated
+    """
+    return isinstance(cls, TypedField) or (
+        isinstance(t.get_origin(cls), type) and issubclass(t.get_origin(cls), TypedField)
+    )
+
+
 ## specific
-def StringField(**kw: Unpack[FieldSettings]) -> TypedField[str]:
+def StringField(**kw: t.Unpack[FieldSettings]) -> TypedField[str]:
     """
     Pydal type is string, Python type is str.
     """
@@ -33,7 +255,7 @@ def StringField(**kw: Unpack[FieldSettings]) -> TypedField[str]:
 String = StringField
 
 
-def TextField(**kw: Unpack[FieldSettings]) -> TypedField[str]:
+def TextField(**kw: t.Unpack[FieldSettings]) -> TypedField[str]:
     """
     Pydal type is text, Python type is str.
     """
@@ -44,7 +266,7 @@ def TextField(**kw: Unpack[FieldSettings]) -> TypedField[str]:
 Text = TextField
 
 
-def BlobField(**kw: Unpack[FieldSettings]) -> TypedField[bytes]:
+def BlobField(**kw: t.Unpack[FieldSettings]) -> TypedField[bytes]:
     """
     Pydal type is blob, Python type is bytes.
     """
@@ -55,7 +277,7 @@ def BlobField(**kw: Unpack[FieldSettings]) -> TypedField[bytes]:
 Blob = BlobField
 
 
-def BooleanField(**kw: Unpack[FieldSettings]) -> TypedField[bool]:
+def BooleanField(**kw: t.Unpack[FieldSettings]) -> TypedField[bool]:
     """
     Pydal type is boolean, Python type is bool.
     """
@@ -66,7 +288,7 @@ def BooleanField(**kw: Unpack[FieldSettings]) -> TypedField[bool]:
 Boolean = BooleanField
 
 
-def IntegerField(**kw: Unpack[FieldSettings]) -> TypedField[int]:
+def IntegerField(**kw: t.Unpack[FieldSettings]) -> TypedField[int]:
     """
     Pydal type is integer, Python type is int.
     """
@@ -77,7 +299,7 @@ def IntegerField(**kw: Unpack[FieldSettings]) -> TypedField[int]:
 Integer = IntegerField
 
 
-def DoubleField(**kw: Unpack[FieldSettings]) -> TypedField[float]:
+def DoubleField(**kw: t.Unpack[FieldSettings]) -> TypedField[float]:
     """
     Pydal type is double, Python type is float.
     """
@@ -88,7 +310,7 @@ def DoubleField(**kw: Unpack[FieldSettings]) -> TypedField[float]:
 Double = DoubleField
 
 
-def DecimalField(n: int, m: int, **kw: Unpack[FieldSettings]) -> TypedField[decimal.Decimal]:
+def DecimalField(n: int, m: int, **kw: t.Unpack[FieldSettings]) -> TypedField[decimal.Decimal]:
     """
     Pydal type is decimal, Python type is Decimal.
     """
@@ -99,7 +321,7 @@ def DecimalField(n: int, m: int, **kw: Unpack[FieldSettings]) -> TypedField[deci
 Decimal = DecimalField
 
 
-def DateField(**kw: Unpack[FieldSettings]) -> TypedField[dt.date]:
+def DateField(**kw: t.Unpack[FieldSettings]) -> TypedField[dt.date]:
     """
     Pydal type is date, Python type is datetime.date.
     """
@@ -110,7 +332,7 @@ def DateField(**kw: Unpack[FieldSettings]) -> TypedField[dt.date]:
 Date = DateField
 
 
-def TimeField(**kw: Unpack[FieldSettings]) -> TypedField[dt.time]:
+def TimeField(**kw: t.Unpack[FieldSettings]) -> TypedField[dt.time]:
     """
     Pydal type is time, Python type is datetime.time.
     """
@@ -121,7 +343,7 @@ def TimeField(**kw: Unpack[FieldSettings]) -> TypedField[dt.time]:
 Time = TimeField
 
 
-def DatetimeField(**kw: Unpack[FieldSettings]) -> TypedField[dt.datetime]:
+def DatetimeField(**kw: t.Unpack[FieldSettings]) -> TypedField[dt.datetime]:
     """
     Pydal type is datetime, Python type is datetime.datetime.
     """
@@ -132,7 +354,7 @@ def DatetimeField(**kw: Unpack[FieldSettings]) -> TypedField[dt.datetime]:
 Datetime = DatetimeField
 
 
-def PasswordField(**kw: Unpack[FieldSettings]) -> TypedField[str]:
+def PasswordField(**kw: t.Unpack[FieldSettings]) -> TypedField[str]:
     """
     Pydal type is password, Python type is str.
     """
@@ -143,7 +365,7 @@ def PasswordField(**kw: Unpack[FieldSettings]) -> TypedField[str]:
 Password = PasswordField
 
 
-def UploadField(**kw: Unpack[FieldSettings]) -> TypedField[str]:
+def UploadField(**kw: t.Unpack[FieldSettings]) -> TypedField[str]:
     """
     Pydal type is upload, Python type is str.
     """
@@ -153,11 +375,10 @@ def UploadField(**kw: Unpack[FieldSettings]) -> TypedField[str]:
 
 Upload = UploadField
 
-T_subclass = typing.TypeVar("T_subclass", TypedTable, Table)
-
 
 def ReferenceField(
-    other_table: str | typing.Type[TypedTable] | TypedTable | Table | T_subclass, **kw: Unpack[FieldSettings]
+    other_table: str | t.Type[TypedTable] | TypedTable | Table | T_subclass,
+    **kw: t.Unpack[FieldSettings],
 ) -> TypedField[int]:
     """
     Pydal type is reference, Python type is int (id).
@@ -180,7 +401,7 @@ def ReferenceField(
 Reference = ReferenceField
 
 
-def ListStringField(**kw: Unpack[FieldSettings]) -> TypedField[list[str]]:
+def ListStringField(**kw: t.Unpack[FieldSettings]) -> TypedField[list[str]]:
     """
     Pydal type is list:string, Python type is list of str.
     """
@@ -191,7 +412,7 @@ def ListStringField(**kw: Unpack[FieldSettings]) -> TypedField[list[str]]:
 ListString = ListStringField
 
 
-def ListIntegerField(**kw: Unpack[FieldSettings]) -> TypedField[list[int]]:
+def ListIntegerField(**kw: t.Unpack[FieldSettings]) -> TypedField[list[int]]:
     """
     Pydal type is list:integer, Python type is list of int.
     """
@@ -202,7 +423,7 @@ def ListIntegerField(**kw: Unpack[FieldSettings]) -> TypedField[list[int]]:
 ListInteger = ListIntegerField
 
 
-def ListReferenceField(other_table: str, **kw: Unpack[FieldSettings]) -> TypedField[list[int]]:
+def ListReferenceField(other_table: str, **kw: t.Unpack[FieldSettings]) -> TypedField[list[int]]:
     """
     Pydal type is list:reference, Python type is list of int (id).
     """
@@ -213,7 +434,7 @@ def ListReferenceField(other_table: str, **kw: Unpack[FieldSettings]) -> TypedFi
 ListReference = ListReferenceField
 
 
-def JSONField(**kw: Unpack[FieldSettings]) -> TypedField[object]:
+def JSONField(**kw: t.Unpack[FieldSettings]) -> TypedField[object]:
     """
     Pydal type is json, Python type is object (can be anything JSON-encodable).
     """
@@ -221,7 +442,7 @@ def JSONField(**kw: Unpack[FieldSettings]) -> TypedField[object]:
     return TypedField(object, **kw)
 
 
-def BigintField(**kw: Unpack[FieldSettings]) -> TypedField[int]:
+def BigintField(**kw: t.Unpack[FieldSettings]) -> TypedField[int]:
     """
     Pydal type is bigint, Python type is int.
     """
@@ -241,7 +462,7 @@ NativeTimestampField = SQLCustomType(
 )
 
 
-def TimestampField(**kw: Unpack[FieldSettings]) -> TypedField[dt.datetime]:
+def TimestampField(**kw: t.Unpack[FieldSettings]) -> TypedField[dt.datetime]:
     """
     Database type is timestamp, Python type is datetime.
 
@@ -275,7 +496,7 @@ def safe_decode_native_point(value: str | None) -> tuple[float, ...]:
 
     try:
         parsed = ast.literal_eval(value)
-        return typing.cast(tuple[float, ...], parsed)
+        return t.cast(tuple[float, ...], parsed)
     except ValueError:  # pragma: no cover
         # should not happen when inserted with `safe_encode_native_point` but you never know
         return ()
@@ -328,7 +549,7 @@ NativePointField = SQLCustomType(
 )
 
 
-def PointField(**kw: Unpack[FieldSettings]) -> TypedField[tuple[float, float]]:
+def PointField(**kw: t.Unpack[FieldSettings]) -> TypedField[tuple[float, float]]:
     """
     Database type is point, Python type is tuple[float, float].
     """
@@ -344,9 +565,14 @@ NativeUUIDField = SQLCustomType(
 )
 
 
-def UUIDField(**kw: Unpack[FieldSettings]) -> TypedField[uuid.UUID]:
+def UUIDField(**kw: t.Unpack[FieldSettings]) -> TypedField[uuid.UUID]:
     """
     Database type is uuid, Python type is UUID.
     """
     kw["type"] = NativeUUIDField
     return TypedField(uuid.UUID, **kw)
+
+
+# note: import at the end to prevent circular imports:
+from .define import TableDefinitionBuilder
+from .tables import TypedTable
